+++ MinVaktServer/src/main/java/minvakt/repos/CategoryRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * MinVakt-Team3
  * 15.01.2017
  */
-public interface CategoryRepository extends JpaRepository<EmployeeCategory, Integer> {
+public interface CategoryRepository extends JpaRepository<EmployeeCategory, Short> {
 
 }
